update regex matching for cjk characters
Regex matching is not kind to non-Latin character sets. Instead of
using word boundaries (\b) for the end of hashtags, I've changed to
a lookahead assertion checking for a space or the end of a line.

I'm going to verify what characters are allowed by twitter in a hashtag
and adjust this lookahead accordingly, in case other characters can break
a hashtag.